Refactored code to have device instance classes and base classes in separate packages/folders
