+++ iot.agile.DeviceFactory/src/main/java/iot/agile/device/base/DeviceImp.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package iot.agile.devicefactory.device;
+package iot.agile.device.base;
 
 import java.util.ArrayList;
 import java.util.HashMap;
