+++ iot.agile.DeviceFactory/src/main/java/iot/agile/device/base/SensorUuid.java
@@ -1,4 +1,4 @@
-package iot.agile.devicefactory.device;
+package iot.agile.device.base;
 
 public class SensorUuid {
 	public String serviceUuid;
