+++ iot.agile.DeviceFactory/src/main/java/iot/agile/device/instance/DummyDevice.java
@@ -1,4 +1,4 @@
-package iot.agile.devicefactory.device;
+package iot.agile.device.instance;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -10,6 +10,7 @@ import org.slf4j.LoggerFactory;
 
 import iot.agile.Device;
 import iot.agile.Protocol;
+import iot.agile.device.base.DeviceImp;
 import iot.agile.exception.AgileNoResultException;
 import iot.agile.object.DeviceComponent;
 import iot.agile.object.DeviceOverview;
