+++ iot.agile.DeviceFactory/src/main/java/iot/agile/device/instance/TISensorTag.java
@@ -1,4 +1,4 @@
-package iot.agile.devicefactory.device;
+package iot.agile.device.instance;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -8,6 +8,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import iot.agile.Device;
+import iot.agile.device.base.AgileBLEDevice;
+import iot.agile.device.base.SensorUuid;
 import iot.agile.exception.AgileNoResultException;
 import iot.agile.object.DeviceDefinition;
 import iot.agile.object.DeviceOverview;
